+++ addon/controllers/google-map/circle.js
@@ -1,0 +1,7 @@
+import Ember from 'ember';
+
+/**
+ * @class GoogleMapCircleController
+ * @extends Ember.ObjectController
+ */
+export default Ember.ObjectController.extend({});
