+++ addon/controllers/google-map/info-window.js
@@ -1,0 +1,7 @@
+import Ember from 'ember';
+
+/**
+ * @class GoogleMapInfoWindowController
+ * @extends Ember.ObjectController
+ */
+export default Ember.ObjectController.extend({});
