+++ addon/controllers/google-map/marker.js
@@ -1,0 +1,7 @@
+import Ember from 'ember';
+
+/**
+ * @class GoogleMapMarkerController
+ * @extends Ember.ObjectController
+ */
+export default Ember.ObjectController.extend({});
