+++ addon/controllers/google-map/markers.js
@@ -1,0 +1,10 @@
+import Ember from 'ember';
+
+/**
+ * @class GoogleMapMarkersController
+ * @extends Ember.ArrayController
+ */
+export default Ember.ArrayController.extend({
+  itemController: Ember.computed.alias('parentController.markerController'),
+  model:          Ember.computed.alias('parentController.markers')
+})
