+++ addon/controllers/google-map/polygon-path.js
@@ -1,0 +1,7 @@
+import GoogleMapPolylinePathController from './polyline-path';
+
+/**
+ * @class GoogleMapPolygonPathController
+ * @extends GoogleMapPolylinePathController
+ */
+export default GoogleMapPolylinePathController.extend({});
