+++ addon/controllers/google-map/polygon.js
@@ -1,0 +1,7 @@
+import GoogleMapPolylineController from './polyline';
+
+/**
+ * @class GoogleMapPolygonController
+ * @extends GoogleMapPolylineController
+ */
+export default GoogleMapPolylineController.extend({});
