+++ app/controllers/google-map/circles.js
@@ -1,10 +1,2 @@
-import Ember from 'ember';
-
-/**
- * @class GoogleMapCirclesController
- * @extends Ember.ArrayController
- */
-export default Ember.ArrayController.extend({
-  itemController: Ember.computed.alias('parentController.circleController'),
-  model:          Ember.computed.alias('parentController.circles')
-});
+import GoogleMapCirclesController from 'ember-google-map/controllers/google-map/circles';
+export default GoogleMapCirclesController;
