+++ app/controllers/google-map/info-windows.js
@@ -1,10 +1,2 @@
-import Ember from 'ember';
-
-/**
- * @class GoogleMapInfoWindowsController
- * @extends Ember.ArrayController
- */
-export default Ember.ArrayController.extend({
-  itemController: Ember.computed.alias('parentController.infoWindowController'),
-  model:          Ember.computed.alias('parentController.infoWindows')
-})
+import GoogleMapInfoWindowsController from 'ember-google-map/controllers/google-map/info-windows';
+export default GoogleMapInfoWindowsController;
