+++ app/controllers/google-map/markers.js
@@ -1,10 +1,2 @@
-import Ember from 'ember';
-
-/**
- * @class GoogleMapMarkersController
- * @extends Ember.ArrayController
- */
-export default Ember.ArrayController.extend({
-  itemController: Ember.computed.alias('parentController.markerController'),
-  model:          Ember.computed.alias('parentController.markers')
-})
+import GoogleMapMarkersController from 'ember-google-map/controllers/google-map/markers';
+export default GoogleMapMarkersController;
