+++ app/controllers/google-map/polygon-path.js
@@ -1,7 +1,2 @@
-import GoogleMapPolylinePathController from './polyline-path';
-
-/**
- * @class GoogleMapPolygonPathController
- * @extends GoogleMapPolylinePathController
- */
-export default GoogleMapPolylinePathController.extend({});
+import GoogleMapPolygonPathController from 'ember-google-map/controllers/google-map/polygon-path';
+export default GoogleMapPolygonPathController;
