+++ app/controllers/google-map/polygon.js
@@ -1,7 +1,2 @@
-import GoogleMapPolylineController from './polyline';
-
-/**
- * @class GoogleMapPolygonController
- * @extends GoogleMapPolylineController
- */
-export default GoogleMapPolylineController.extend({});
+import GoogleMapPolygonController from 'ember-google-map/controllers/google-map/polygon';
+export default GoogleMapPolygonController;
