+++ app/controllers/google-map/polyline.js
@@ -1,15 +1,2 @@
-import Ember from 'ember';
-
-/**
- * @class GoogleMapPolylineController
- * @extends Ember.ObjectController
- */
-export default Ember.ObjectController.extend({
-  pathController: Ember.computed.alias('parentController.pathController'),
-
-  _path: Ember.computed('path', 'pathController', function () {
-    return this.container.lookupFactory('controller:' + this.get('pathController')).create({
-      parentController: this
-    });
-  }).readOnly()
-});
+import GoogleMapPolylineController from 'ember-google-map/controllers/google-map/polyline';
+export default GoogleMapPolylineController;
